Transition Razor test comparers to new project.
- Changed namespaces for test comparers.

#461
