+++ src/Microsoft.AspNet.Razor.Test.Sources/project.json
@@ -1,0 +1,13 @@
+﻿{
+  "version": "1.0.0-*",
+  "shared": "*.cs",
+  "dependencies": { },
+  "frameworks": {
+    "net45": { },
+    "dotnet": {
+      "dependencies": {
+        "System.Runtime": "4.0.20-*"
+      }
+    }
+  }
+}
